Move loadObjectReferences to proxy-db
