+++ packages/core/echo/echo-db/src/proxy-db/index.ts
@@ -4,3 +4,4 @@
 
 export * from './database';
 export * from './dynamic-schema-registry';
+export * from './load-object-references';
